Introduce base class for object can be saved
